pyproject.toml
[tool.poetry]
name = "dogs-django"
version = "0.1.0"
description = "dogs"
authors = ["abyssal <abyssal>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
djangorestframework = "^3.15.2"
django = "^5.1.7"
dotenv = "^0.9.9"
psycopg2-binary = "^2.9.10"
drf-yasg = "^1.21.10"


[tool.poetry.group.dev.dependencies]
ruff = "^0.11.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 119

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
ignore = ["D107"] # ignore missing __init__ docstrings

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"